requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

numpy==1.24.2
pandas==2.0.0
opencv-python-headless==4.7.0.72
tqdm==4.65.0
scikit-learn==1.2.2
omegaconf==2.3.0

torch==2.0.0+cpu
torchvision==0.15.1
pytorch-metric-learning==2.1.0
tensorboard==2.12.2
timm==0.6.13

kaggle==1.5.13
dvc==2.55.0
dvc-gdrive==2.19.2

black==23.3.0
pytest==7.3.2
pytest-black==0.3.12
pytest-cov==4.1.0

pymongo==4.3.3
ansible-vault==2.1.0
kafka-python==2.0.2
